Refactor project structure; remove unused components and styles, update AboutMe and Projects sections
