+++ src/Components/Card.jsx
@@ -1,16 +1,14 @@
 import "./Card.css"
 
-export const Card = ({ title, text, icon }) => {
+
+export const Card = ({ name, discription, icon }) => {
   return (
     <article className="card">
       <img src={icon} alt="" className="card-image" />
-      <h2>{title}</h2>
-      <p>{text}</p>
-      <div> <OrangeButtons /> <Lightbuttons />
-      </div>
+      <h2>{name}</h2>
+      <p>{discription}</p>
+      
     </article>
-  )
-}
-
-
+  );
+};
