+++ src/Components/GithubButton.css
@@ -1,0 +1,22 @@
+a {
+  border-radius: 12px;
+  box-shadow: var(--shadow);
+  cursor: pointer;
+  font-family: "IBM Plex Mono", sans-serif;
+  font-size: 1rem;
+  font-weight: 600;
+  padding: 1rem 2rem;
+  text-align: center;
+  background: var(--color-accent);
+  color: var(--font-color-primary);
+  border: solid 2px var(--color-accent);
+}
+a:hover {
+  background: var(--color-nuetral);
+  color: var(--font-color-secondary);
+  border: solid 2px var(--color-primary);
+}
+a:active {
+border: transparent;
+  box-shadow: inset var(--shadow);
+}
