+++ src/Components/NetlifyButton.css
@@ -1,0 +1,21 @@
+a {
+  border-radius: 12px;
+  box-shadow: var(--shadow);
+  cursor: pointer;
+  font-family: "IBM Plex Mono", sans-serif;
+  font-size: 1rem;
+  font-weight: 600;
+  padding: 1rem 2rem;
+  text-align: center;
+  background: var(--color-neutral);
+  color: var(--font-color-primary);
+  border: solid 2px var(--color-primary);
+}
+a:hover {
+  background: var(--color-primary);
+  color: var(--font-color-secondary);
+}
+a:active {
+border: transparent;
+  box-shadow: inset var(--shadow);
+}
