+++ src/Sections/Head.jsx
@@ -1,4 +1,4 @@
-import { Navbar } from "../../Components/Navbar/Navbar";
+import { Navbar } from "../Components/Navbar";
 
 
 export const Header = () => {
